move mutation to subpackage pkg/engine/mutate
