+++ pkg/engine/utils_test.go
@@ -1,32 +1,13 @@
 package engine
 
 import (
-	"encoding/json"
 	"testing"
 
 	kyverno "github.com/nirmata/kyverno/pkg/api/kyverno/v1"
+	"github.com/nirmata/kyverno/pkg/engine/utils"
 	"gotest.tools/assert"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
-
-func TestGetAnchorsFromMap_ThereAreNoAnchors(t *testing.T) {
-	rawMap := []byte(`{
-		"name":"nirmata-*",
-		"notAnchor1":123,
-		"namespace":"kube-?olicy",
-		"notAnchor2":"sample-text",
-		"object":{
-			"key1":"value1",
-			"(key2)":"value2"
-		}
-	}`)
-
-	var unmarshalled map[string]interface{}
-	json.Unmarshal(rawMap, &unmarshalled)
-
-	actualMap := getAnchorsFromMap(unmarshalled)
-	assert.Assert(t, len(actualMap) == 0)
-}
 
 // Match multiple kinds
 func TestResourceDescriptionMatch_MultipleKind(t *testing.T) {
@@ -68,7 +49,7 @@ func TestResourceDescriptionMatch_MultipleKind(t *testing.T) {
 		   }
 		}
 	 }`)
-	resource, err := ConvertToUnstructured(rawResource)
+	resource, err := utils.ConvertToUnstructured(rawResource)
 	if err != nil {
 		t.Errorf("unable to convert raw resource to unstructured: %v", err)
 
@@ -125,7 +106,7 @@ func TestResourceDescriptionMatch_Name(t *testing.T) {
 		   }
 		}
 	 }`)
-	resource, err := ConvertToUnstructured(rawResource)
+	resource, err := utils.ConvertToUnstructured(rawResource)
 	if err != nil {
 		t.Errorf("unable to convert raw resource to unstructured: %v", err)
 
@@ -183,7 +164,7 @@ func TestResourceDescriptionMatch_Name_Regex(t *testing.T) {
 		   }
 		}
 	 }`)
-	resource, err := ConvertToUnstructured(rawResource)
+	resource, err := utils.ConvertToUnstructured(rawResource)
 	if err != nil {
 		t.Errorf("unable to convert raw resource to unstructured: %v", err)
 
@@ -241,7 +222,7 @@ func TestResourceDescriptionMatch_Label_Expression_NotMatch(t *testing.T) {
 		   }
 		}
 	 }`)
-	resource, err := ConvertToUnstructured(rawResource)
+	resource, err := utils.ConvertToUnstructured(rawResource)
 	if err != nil {
 		t.Errorf("unable to convert raw resource to unstructured: %v", err)
 
@@ -307,7 +288,7 @@ func TestResourceDescriptionMatch_Label_Expression_Match(t *testing.T) {
 		   }
 		}
 	 }`)
-	resource, err := ConvertToUnstructured(rawResource)
+	resource, err := utils.ConvertToUnstructured(rawResource)
 	if err != nil {
 		t.Errorf("unable to convert raw resource to unstructured: %v", err)
 
@@ -375,7 +356,7 @@ func TestResourceDescriptionExclude_Label_Expression_Match(t *testing.T) {
 		   }
 		}
 	 }`)
-	resource, err := ConvertToUnstructured(rawResource)
+	resource, err := utils.ConvertToUnstructured(rawResource)
 	if err != nil {
 		t.Errorf("unable to convert raw resource to unstructured: %v", err)
 
@@ -410,16 +391,4 @@ func TestResourceDescriptionExclude_Label_Expression_Match(t *testing.T) {
 		ExcludeResources: kyverno.ExcludeResources{ResourceDescription: resourceDescriptionExclude}}
 
 	assert.Assert(t, !MatchesResourceDescription(*resource, rule))
-}
-
-func TestRemoveAnchor_ConditionAnchor(t *testing.T) {
-	assert.Equal(t, removeAnchor("(abc)"), "abc")
-}
-
-func TestRemoveAnchor_ExistanceAnchor(t *testing.T) {
-	assert.Equal(t, removeAnchor("^(abc)"), "abc")
-}
-
-func TestRemoveAnchor_EmptyExistanceAnchor(t *testing.T) {
-	assert.Equal(t, removeAnchor("^()"), "")
 }
