+++ pkg/webhooks/policymutation_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	kyverno "github.com/nirmata/kyverno/pkg/api/kyverno/v1"
-	"github.com/nirmata/kyverno/pkg/engine"
+	"github.com/nirmata/kyverno/pkg/engine/utils"
 	"gotest.tools/assert"
 )
 
@@ -31,7 +31,7 @@ func TestGeneratePodControllerRule_NilAnnotation(t *testing.T) {
 	patches, errs := generatePodControllerRule(policy)
 	assert.Assert(t, len(errs) == 0)
 
-	p, err := engine.ApplyPatches(policyRaw, patches)
+	p, err := utils.ApplyPatches(policyRaw, patches)
 	assert.NilError(t, err)
 
 	expectedPolicy := []byte(`{
@@ -83,7 +83,7 @@ func TestGeneratePodControllerRule_ExistOtherAnnotation(t *testing.T) {
 	patches, errs := generatePodControllerRule(policy)
 	assert.Assert(t, len(errs) == 0)
 
-	p, err := engine.ApplyPatches(policyRaw, patches)
+	p, err := utils.ApplyPatches(policyRaw, patches)
 	assert.NilError(t, err)
 
 	expectedPolicy := []byte(`{
@@ -233,7 +233,7 @@ func TestGeneratePodControllerRule(t *testing.T) {
 	patches, errs := generatePodControllerRule(policy)
 	assert.Assert(t, len(errs) == 0)
 
-	p, err := engine.ApplyPatches(policyRaw, patches)
+	p, err := utils.ApplyPatches(policyRaw, patches)
 	assert.NilError(t, err)
 
 	expectPolicy := []byte(`{
